Move port management code into ports directory. NFC
This change splits the logically seperate port building code out of
system_libs.py.
